+++ CHANGES.md
@@ -9,6 +9,7 @@
     do not inject `Object` on full IDL tests
 - Testing (W3C): Add `structured-clone.any.js` to known `badFiles` list
 - Testing (W3C): Add `MessageChannel`
+- Testing (W3C): Add canvas for `ImageData`
 - npm: Bump typeson-registry (Cause `Event` and `MessageChannel` to
     throw upon attempted cloning)